fix : 패키지 구조 변경 및 Dto JavaDoc 주석
