+++ src/main/java/damon/backend/dto/response/calendar/CalendarsResponseDto.java
@@ -1,4 +1,4 @@
-package damon.backend.dto.response;
+package damon.backend.dto.response.calendar;
 
 import damon.backend.entity.Area;
 import damon.backend.entity.Calendar;
@@ -13,21 +13,24 @@ import java.util.List;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
-public class CalendarResponseDto {
+public class CalendarsResponseDto {
+    private Long calendarId;
     private String title;
     private LocalDate startDate;
     private LocalDate endDate;
     private Area area;
 
-    private List<TravelDetailDto> travels;
-
-    public static CalendarResponseDto from(Calendar calendar) {
-        return new CalendarResponseDto(
+    public static CalendarsResponseDto from(Calendar calendar) {
+        return new CalendarsResponseDto(
+                calendar.getId(),
                 calendar.getTitle(),
                 calendar.getStartDate(),
                 calendar.getEndDate(),
-                calendar.getArea(),
-                TravelDetailDto.listFrom(calendar.getTravels())
+                calendar.getArea()
         );
+    }
+
+    public static List<CalendarsResponseDto> listFrom(List<Calendar> calendars) {
+        return calendars.stream().map(CalendarsResponseDto::from).toList();
     }
 }
